requirements.txt
# Core
streamlit
fastapi
requests
python-dotenv

# Audio Processing

ffmpeg-python

# IBM Watson SDKs
ibm-watson
ibm-watson-machine-learning

# Optional / Helper (for future phases)
PyMuPDF     # for PDF extraction
SpeechRecognition  # for voice commands (Phase 3+)

streamlit
python-dotenv
requests

pydub
huggingface-hub


# Text-to-speech
gTTS>=2.5.1
pydub>=0.25.1

# Optional: if you want Groq's official SDK (not required, we use requests)
groq>=0.9.0

google-genai>=0.13.0